Добавил сетевые запросы для категорий, методы для базы данных, юзкейсы для них.

+++ app/src/main/java/com/example/newstoday/data/local/entity/ArticleDBO.kt
@@ -9,7 +9,7 @@ import com.google.gson.Gson
 
 @Entity
 data class ArticleDBO(
-    @PrimaryKey(autoGenerate = true) val id: Long,
+    @PrimaryKey(autoGenerate = true) val id: Int,
     @ColumnInfo("source") val source: Source,
     @ColumnInfo("author") val author: String,
     @ColumnInfo("title") val title: String,
